[FIX] hw_drivers: Don't terminate Manager thread on exception
If you unplug the iot box and wait 15 seconds or so, and then plug it
back in, it's possible to get an exception from helpers.get_ip, because
neither the wired nor the wireless interface has an IP. In that case,
the exception would kill the Manager thread, since it was uncaught.

This fix adds a catch-all to the Manager thread that logs any uncaught
exceptions and continues its main loop.

closes odoo/odoo#77106

X-original-commit: 8beb6252f2d684cc32d12c22c7d8dbb46212b192
Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>
Signed-off-by: Raf Geens <raf-odoo@users.noreply.github.com>