[FIX] web: scrolling: take html to scroll in mobile
Before this commit the scroll position in mobile mode were not kept.
This was because the scroll element was the body node.

After this commit, the scroll is saved and reused correctly as the reference
node is the html node.

This commit makes no sense without its enterprise counterpart. (https://github.com/odoo/enterprise/pull/21150)

closes odoo/odoo#77087

Related: odoo/enterprise#21150
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>