[FIX] website: prevent snippet drop in product_catalog
Before this commit, a user would be able to drop pretty much any
snippet inside the product_catalog (pricelist) snippet.
This was not ideal as the snippet wasn't designed to receive
any other snippets inside of it.

With this commit, the configuration of the snippet is closer
to snippets such as timeline, allowing lines to be moved around
without allowing snippets to be dropped inside it.

task-2655171

closes odoo/odoo#78178

X-original-commit: 0b74eb2ad4ad3a65bc278c3a6367a6e5d1ae85e3
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>