[FIX] website: fix non-full layouts combined with footer scroll effect
Since [1], a footer effect option has been introduced which allows to
have the content hovering the footer to then reveal it when reaching the
bottom of the page (instead of having the footer to scroll like the rest
of the content).

To achieve that, the body color was forced on the snippet container (the
`<main/>`) so that the transparent snippet still appear with a
background color for the time they go over the footer.
That background color should however not have been forced in non-full
layouts as another is already forced in that case. In those non-full
layout, the "full" body background (= background of color combination 1)
is applied on the snippet container but the "body" color is another
color which is customizable by the user too.

[1]: https://github.com/odoo/odoo/commit/6dbe38a3930558609154362e1a343b505ee99b5d

Related to task-2590182

closes odoo/odoo#77517

X-original-commit: fb2cd22bb79f0622878e59dff7527451e6c5a08f
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>