[FIX] website_sale: restore the block ui loading on pay now click
The cart UX was improved with [1] at payment step by adding a block ui when the
user had clicked on "Pay now".
Indeed, that step could take some (long) time, and preventing the user to click
on anything would limit the possible issues: changing a delivery, a payment
provider or anything.
Also, it would be even clearer to the user that something is being processed,
even if the Pay now button was already spinning.

The payment refactoring [2] did not correctly adapt that behavior.
This commit restores it.

[1]: https://github.com/odoo/odoo/commit/0630571ec31b74ae866cb1e2e3d4609cf5fc3b82
[2]: https://github.com/odoo/odoo/commit/5639960a7eb88f1f7b1aae0754905ad9fd9c8556

task-2659450

closes odoo/odoo#78300

X-original-commit: 316915da7c3995dd6c1c94c299fee49de212f18d
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>
Signed-off-by: Romain Derie (rde) <rde@odoo.com>