[FIX] account: use refund repartition line when needed when applying a reconciliation model with a tax from a button
The fix made in https://github.com/odoo/odoo/pull/74427 was mistakenly not forward-ported. We reintroduce it here together with the fix that was made for it by https://github.com/odoo/odoo/pull/77589 .

Before this fix, the wrong repartition was used, and the tax_tag_invert computed for the move lines in this case was inconsistent, leading to wrong reporting.

closes odoo/odoo#78470

X-original-commit: 35f062ef2183debe5d85696d7f582152296c8e41
Related: odoo/enterprise#21733
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>