[FIX] project: improve performances of has any so to invoice compute
Before this commit, the compute _compute_has_any_so_to_invoice was not triggered
in case of user change, which means that it could take incorrect values depending
on the value of the group as it was missing the proper decorator.
Also, performances were not optimal and have been improved.

closes odoo/odoo#77590

Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>