[FIX] web_editor: restore column count snippet option
When decreasing the number of columns through the column count snippet
option, the editor entered a deadlock situation.

This reverts [1] which is the cause, trying to await the UI update
during a snippet removal. This commit first prevents the deadlock while
waiting for a better solution to be implemented.

[1]: https://github.com/odoo/odoo/commit/167e3c6119ae766b67d998fdabd25f1f74a573a5

task-2652904

closes odoo/odoo#77476

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>