[IMP] website: use initAutoMoreMenu for website menu
The goal of the 'autohideMenu' feature: instead of having the menu
items wrapped on multiple lines, they are hidden and only when we
compute how many items we can show on one line, we show the menu.

This implementation was fine in old versions when we did not have
the lazy loading of the JS yet.

The goal of this commit is to replace 'initAutoMoreMenu' by code
that can be called from a non-lazy loaded JS file so it will be
possible to remove fade-in effect from it.

task-2618929

Part-of: odoo/odoo#75063